git-svn-id: https://taverna.googlecode.com/svn/taverna/ui/net.sf.taverna.t2.ui-components/branches/ui-components-1.3@13485 bf327186-88b3-11dd-a302-d386e5130c1c

+++ run-ui/src/main/java/net/sf/taverna/t2/workbench/run/cleanup/DatabaseCleanup.java
@@ -99,6 +99,9 @@ public class DatabaseCleanup {
 	public void scheduleDeleteDataflowRun(String workflowRunId,
 			boolean startDeletion) {
 		addToDeletionQueue(workflowRunId, startDeletion);
+		synchronized (deletionOnRestartList) {
+			deletionOnRestartList.remove(workflowRunId);	
+		}
 	}
 	
 	public void scheduleDeleteDataflowRunOnRestart(String workflowRunId) {
@@ -138,7 +141,7 @@ public class DatabaseCleanup {
 
 	protected void persist() {
 		File runToDeleteFile = getRunToDeleteFile();
-		if (deletionQueue.isEmpty()) {
+		if (deletionQueue.isEmpty() && deletionOnRestartList.isEmpty()) {
 			runToDeleteFile.delete();
 			return;
 		}
